feat(governance): engine moves to the kernel with injected defaults (KJC-TSK-0745)
